refactor: change manual -> versioning for readability
